+++ sda/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- ingest: Do not keep a transaction open while writing to the archive storage
+  - Separate transaction that is commited before writing to the archive if file is now known in db but present in inbox storage 
+  - Write the "submitted" file event after writing to the archive
+  - Delete the file from the archive is db actions fail
+
 ## [4.0.0] - 2026-09-21
 
 ### Added